chore: Moved validation into common module
