+++ ra2ce/common/validation/ra2ce_validator_protocol.py
@@ -22,7 +22,7 @@
 
 from typing import Protocol
 
-from ra2ce.validation.validation_report import ValidationReport
+from ra2ce.common.validation.validation_report import ValidationReport
 
 
 class Ra2ceIoValidator(Protocol):
