+++ ra2ce/configuration/validators/ini_config_validator_base.py
@@ -24,8 +24,8 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 from ra2ce.configuration.ini_config_protocol import IniConfigDataProtocol
-from ra2ce.validation.ra2ce_validator_protocol import Ra2ceIoValidator
-from ra2ce.validation.validation_report import ValidationReport
+from ra2ce.common.validation.ra2ce_validator_protocol import Ra2ceIoValidator
+from ra2ce.common.validation.validation_report import ValidationReport
 
 IndirectAnalysisNameList: List[str] = [
     "single_link_redundancy",
