+++ tests/configuration/test_ini_config_validator_base.py
@@ -7,7 +7,7 @@ from ra2ce.configuration.ini_config_protocol import IniConfigDataProtocol
 from ra2ce.configuration.validators.ini_config_validator_base import (
     IniConfigValidatorBase,
 )
-from ra2ce.validation.validation_report import ValidationReport
+from ra2ce.common.validation.validation_report import ValidationReport
 from tests import test_data, test_results
